+++ lib/liberic/boot.rb
@@ -19,7 +19,7 @@ module Liberic
     File.expand_path("libericapi.#{suffix}", ERIC_LIB_FOLDER)
   end
 
-  ERIC_HOME = ENV['ERIC_HOME_39'] || raise(InitializationError.new('ERIC_HOME_39 environment variable not found (set it to the path to the ERiC libraries)'))
+  ERIC_HOME = ENV['ERIC_HOME_40'] || raise(InitializationError.new('ERIC_HOME_40 environment variable not found (set it to the path to the ERiC libraries)'))
   ERIC_LIB_FOLDER = File.expand_path('lib', ERIC_HOME)
 
   def check_eric_version!